добавил подходы в сессию + защитил все юзкейсы проверками экзист на юзера/др. сущности , спасая от эрор500

+++ src/main/java/com/example/fitness_assistant/core/model/Set.java
@@ -15,7 +15,7 @@ public class Set {
     private Long id;
     private Long sessionId;
     private Long exerciseId;
-    private Long weight;
+    private Double weight;
     private Integer reps;
     private LocalDateTime createdAt;
 }
